+++ pylabrobot/storage/__init__.py
@@ -4,3 +4,6 @@ from .cytomat import CytomatBackend
 from .incubator import Incubator
 from .inheco.scila import SCILABackend
 from .liconic import ExperimentalLiconicBackend
+from .stacker import EmptyStackError, LoadingTrayOccupiedError, Stacker
+from .stacker_backend import StackerBackend
+from .stacker_chatterbox import StackerChatterboxBackend